Migrate from Next.js pages router to app router
- Update dependencies to Next.js 14 and React 18
- Migrate all pages from pages/ to app/ directory structure
- Convert _app.js and _document.js to app/layout.js
- Add proper client/server component boundaries with 'use client'
- Update routing from next/router to next/navigation
- Convert getStaticProps to async server components
- Maintain all existing functionality and routes
