+++ lib/search-context.js
@@ -1,3 +1,5 @@
+'use client'
+
 import { createContext, useContext } from 'react'
 
 export const SearchContext = createContext(null)
