+++ next.config.js
@@ -1,7 +1,11 @@
 const withMDX = require('@next/mdx')({
   extension: /\.mdx?$/,
+  options: {
+    remarkPlugins: [],
+    rehypePlugins: [],
+  },
 })
 
 module.exports = withMDX({
-  pageExtensions: ['js', 'mdx'],
+  pageExtensions: ['js', 'jsx', 'mdx', 'ts', 'tsx'],
 })
